WIP add new component for menu
