+++ src/components/calcite-action-group/calcite-action-group.scss
@@ -14,15 +14,3 @@
   display: flex;
   width: 100%;
 }
-
-.menu-button {
-  align-self: stretch;
-  flex: 0 1 auto;
-  height: 100%;
-  position: relative;
-}
-
-.menu {
-  min-width: var(--calcite-menu-min-width);
-  flex-flow: column nowrap;
-}
