+++ src/components/calcite-action-group/calcite-action-group.tsx
@@ -1,10 +1,6 @@
-import { Component, Host, h, Element, Prop, Watch } from "@stencil/core";
-import { CSS, ICONS, SLOTS, TEXT } from "./resources";
-import { focusElement, getSlotted } from "../../utils/dom";
+import { Component, Host, h, Prop, Watch } from "@stencil/core";
+import { SLOTS } from "./resources";
 import { VNode } from "@stencil/core/internal";
-import { getRoundRobinIndex } from "../../utils/array";
-
-const SUPPORTED_ARROW_KEYS = ["ArrowUp", "ArrowDown"];
 
 @Component({
   tag: "calcite-action-group",
@@ -54,82 +50,23 @@ export class CalciteActionGroup {
 
   // --------------------------------------------------------------------------
   //
-  //  Private Properties
-  //
-  // --------------------------------------------------------------------------
-
-  @Element() el: HTMLCalciteActionGroupElement;
-
-  menuButtonEl: HTMLCalciteActionElement;
-
-  // --------------------------------------------------------------------------
-  //
   //  Component Methods
   //
   // --------------------------------------------------------------------------
 
-  renderMenuButton(): VNode {
-    const { menuOpen, intlOpen, intlOptions, intlClose, expanded } = this;
-    const closeLabel = intlClose || TEXT.close;
-    const openLabel = intlOpen || TEXT.open;
-    const optionsText = intlOptions || TEXT.options;
-    const menuLabel = menuOpen ? closeLabel : openLabel;
-
-    return (
-      <calcite-action
-        active={menuOpen}
-        aria-label={menuLabel}
-        class={CSS.menuButton}
-        icon={ICONS.menu}
-        label={menuLabel}
-        onClick={this.toggleMenuOpen}
-        onKeyDown={this.menuButtonKeyDown}
-        ref={this.setMenuButonRef}
-        text={optionsText}
-        textEnabled={expanded}
-      />
-    );
-  }
-
   renderMenu(): VNode {
-    const { el } = this;
-
-    const actionCount = getSlotted(el, SLOTS.menuActions, { all: true }).length;
-    const hasMin = actionCount > 1;
-
-    if (actionCount && !hasMin) {
-      console.warn(`${el.tagName}: '${SLOTS.menuActions}' slot must have a minimum of 2 actions.`, {
-        el
-      });
-    }
-
-    return hasMin ? (
-      <div class={CSS.menuContainer} onKeyDown={this.menuActionsContainerKeyDown}>
-        {this.renderMenuButton()}
-        {this.renderMenuItems()}
-      </div>
-    ) : null;
-  }
-
-  renderMenuItems(): VNode {
-    const { menuOpen, menuButtonEl, intlOptions } = this;
-    const label = intlOptions || TEXT.options;
+    const { expanded, intlClose, intlOpen, intlOptions, menuOpen } = this;
 
     return (
-      <calcite-popover
-        disablePointer={true}
-        flipPlacements={["left", "right"]}
-        label={label}
-        offsetDistance={8}
-        onKeyDown={this.menuActionsKeydown}
+      <calcite-action-menu
+        expanded={expanded}
+        intlClose={intlClose}
+        intlOpen={intlOpen}
+        intlOptions={intlOptions}
         open={menuOpen}
-        placement="leading"
-        referenceElement={menuButtonEl}
       >
-        <div class={CSS.menu}>
-          <slot name={SLOTS.menuActions} />
-        </div>
-      </calcite-popover>
+        <slot name={SLOTS.menuActions} />
+      </calcite-action-menu>
     );
   }
 
@@ -141,98 +78,4 @@ export class CalciteActionGroup {
       </Host>
     );
   }
-
-  // --------------------------------------------------------------------------
-  //
-  //  Private Methods
-  //
-  // --------------------------------------------------------------------------
-
-  setMenuButonRef = (node: HTMLCalciteActionElement): void => {
-    this.menuButtonEl = node;
-  };
-
-  queryActions(): HTMLCalciteActionElement[] {
-    return getSlotted<HTMLCalciteActionElement>(this.el, SLOTS.menuActions, {
-      all: true
-    });
-  }
-
-  isValidKey(key: string, supportedKeys: string[]): boolean {
-    return !!supportedKeys.find((k) => k === key);
-  }
-
-  menuActionsKeydown = (event: KeyboardEvent): void => {
-    const { key, target } = event;
-
-    if (!this.isValidKey(key, SUPPORTED_ARROW_KEYS)) {
-      return;
-    }
-
-    const actions = this.queryActions();
-    const { length } = actions;
-    const currentIndex = actions.indexOf(target as HTMLCalciteActionElement);
-
-    if (!length || currentIndex === -1) {
-      return;
-    }
-
-    event.preventDefault();
-
-    if (key === "ArrowUp") {
-      const value = getRoundRobinIndex(currentIndex - 1, length);
-      const previousAction = actions[value];
-      focusElement(previousAction);
-    }
-
-    if (key === "ArrowDown") {
-      const value = getRoundRobinIndex(currentIndex + 1, length);
-      const nextAction = actions[value];
-      focusElement(nextAction);
-    }
-  };
-
-  menuActionsContainerKeyDown = (event: KeyboardEvent): void => {
-    const { key } = event;
-
-    if (key === "Escape") {
-      this.menuOpen = false;
-    }
-  };
-
-  toggleMenuOpen = (): void => {
-    this.menuOpen = !this.menuOpen;
-  };
-
-  menuButtonKeyDown = (event: KeyboardEvent): void => {
-    const { key } = event;
-    const { menuOpen } = this;
-
-    if (!this.isValidKey(key, SUPPORTED_ARROW_KEYS)) {
-      return;
-    }
-
-    const actions = this.queryActions();
-    const { length } = actions;
-
-    if (!length) {
-      return;
-    }
-
-    event.preventDefault();
-
-    if (!menuOpen) {
-      this.menuOpen = true;
-    }
-
-    if (key === "ArrowUp") {
-      const lastAction = actions[length - 1];
-      focusElement(lastAction);
-    }
-
-    if (key === "ArrowDown") {
-      const firstAction = actions[0];
-      focusElement(firstAction);
-    }
-  };
 }
