+++ src/components/calcite-action-group/resources.ts
@@ -1,20 +1,3 @@
-export const CSS = {
-  menuContainer: "menu-container",
-  menuButton: "menu-button",
-  menu: "menu",
-  menuOpen: "menu--open"
-};
-
-export const ICONS = {
-  menu: "ellipsis"
-};
-
 export const SLOTS = {
   menuActions: "menu-actions"
-};
-
-export const TEXT = {
-  close: "Close",
-  open: "Open",
-  options: "Options"
 };
