+++ src/components/calcite-action-menu/resources.ts
@@ -1,0 +1,15 @@
+export const CSS = {
+  menuContainer: "menu-container",
+  menuButton: "menu-button",
+  menu: "menu"
+};
+
+export const ICONS = {
+  menu: "ellipsis"
+};
+
+export const TEXT = {
+  close: "Close",
+  open: "Open",
+  options: "Options"
+};
